Make internal modules private
We don't really want this package to be used as a library.
To properly show that, make our modules private.

With this, the public api is the main function and the kernel
provisioner class but I would not recommend relying on implementation
details of them either.
