+++ src/pyproject_local_kernel/_configdata.py
